Millisecond data merge + debug
Attempts to find data discrepancies are half-way, even in rainfall seems now o.k., soil water is inconsistent with it. Rain total used totalize in the logger and this seems to explain the error, but soil water used Average and shows an error by the same factor... problem remains to be solved!